+++ app/controllers/api/organizations_controller.rb
@@ -3,10 +3,11 @@
 module Api
   class OrganizationsController < ApiController
     has_scope :exclude_deleted, type: :boolean
-    has_scope :after_timestamp
+    # has_scope :after_timestamp: we ignore the after_timestamp scope for organizations API. We want to always return all organizations the client has access to
+    after_action :verify_policy_scoped, only: :index
 
     def index
-      @organizations = apply_scopes(Organization).all
+      @organizations = apply_scopes(policy_scope(Organization)).all
       respond_with @organizations, include: included_params, meta: { timestamp: Time.zone.now }
     end