+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 - Upgraded capybara from 2.17.0 to 2.18.0
 - Upgraded i18n from 0.9.4 to 0.9.5
+- Applying Policy scope to Organizations API
+- ignoring after_timestamp parameter for Organizations API
 
 ## 0.1.0
 - Base Release